requirements.txt
allure-pytest==2.14.1
allure-python-commons==2.14.1
attrs==25.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.5